update hqt source code
