+++ neural_compressor/torch/algorithms/fp8_quant/habana_quantization_toolkit/README.md
@@ -1,0 +1,4 @@
+# The Quantization Toolkit
+
+The Quantization Toolkit (HQT) provides model measurement and quantization capabilities in PyTorch.
+For full details see: https://docs.habana.ai/en/latest/PyTorch/Inference_on_PyTorch/Inference_Using_FP8.html
