+++ neural_compressor/torch/algorithms/fp8_quant/habana_quantization_toolkit/_core/scale_methods/__init__.py
