+++ neural_compressor/torch/algorithms/fp8_quant/habana_quantization_toolkit/custom_config/custom_example.json
@@ -1,0 +1,5 @@
+{
+    "mode": "MEASURE",
+    "scale_method": "MAX",
+    "fp8_config": "E4M3" 
+}
