+++ neural_compressor/torch/algorithms/fp8_quant/habana_quantization_toolkit/custom_config/llama_measure.json
@@ -1,0 +1,8 @@
+{
+    "mode": "MEASURE",
+    "observer": "maxabs",
+    "allowlist": {"types": [], "names":  []},
+    "blocklist": {"types": [], "names":  []},
+    "quantize_weight": false,
+    "dump_stats_path": "./llama_output/7b_measure"
+}
