+++ neural_compressor/torch/algorithms/fp8_quant/habana_quantization_toolkit/custom_config/llama_quant.json
@@ -1,0 +1,9 @@
+{
+    "mode": "QUANTIZE",
+    "observer": "maxabs",
+    "scale_method": "maxabs_hw",
+    "allowlist": {"types": [], "names":  []},
+    "blocklist": {"types": [], "names":  ["lm_head"]},
+    "quantize_weight": false,
+    "dump_stats_path": "./llama_output/7b_measure"
+}
