+++ neural_compressor/torch/algorithms/fp8_quant/habana_quantization_toolkit/custom_config/measure_config.json
@@ -1,0 +1,12 @@
+{
+    "mode": "MEASURE",
+    "scale_method": "MAX",
+    "quantize_weight": true,
+    "dump_stats_path": "./run_outputs/fp8/stats",
+    "allowlist": {
+        "types": [
+            "torch.nn.Linear",
+            "torch.nn.Conv2d"
+        ]
+    }
+}
