+++ neural_compressor/torch/algorithms/fp8_quant/habana_quantization_toolkit/custom_config/quant_config.json
@@ -1,0 +1,12 @@
+{
+    "mode": "QUANTIZE",
+    "scale_method": "without_scale",
+    "fp8_config": "E4M3",
+    "allowlist": {
+        "types": [
+            "torch.nn.Linear",
+            "torch.nn.Conv2d"
+        ]
+    },
+    "dump_stats_path": "./run_outputs/fp8/stats"
+}
