+++ neural_compressor/torch/algorithms/fp8_quant/habana_quantization_toolkit/tests/test_jsons/test_measure.json
@@ -1,0 +1,7 @@
+{
+    "mode": "MEASURE",
+    "observer": "maxabs",
+    "allowlist": {"types": [], "names":  []},
+    "blocklist": {"types": [], "names":  []},
+    "dump_stats_path": "./test_outputs/unit_test"
+}
