+++ neural_compressor/torch/algorithms/fp8_quant/habana_quantization_toolkit/tests/test_jsons/test_pow2_quant.json
@@ -1,0 +1,8 @@
+{
+    "mode": "QUANTIZE",
+    "observer": "maxabs",
+    "scale_method": "maxabs_pow2",
+    "allowlist": {"types": [], "names":  []},
+    "blocklist": {"types": [], "names":  ["lm_head"]},
+    "dump_stats_path": "./test_outputs/unit_test"
+}
