--- neural_compressor/torch/algorithms/habana_fp8/quantization_toolkit/_module_method/__init__.py
@@ -1,2 +1,0 @@
-from .modules import QuantizedLinear, QuantizedConv2d, QuantizedModule, QuantizedBmm
-from .quant_util import get_named_modules
